Transfering and Clearing the Spaceship Buffer
Got the assembly functions done for transferring the 16kb spaceship buffer to VRAM, and for Clearing the buffer afterwards.

Together, they take up around 1/10th of a frame's CPU time. That makes me nervous about how long it will take to draw the spaceship, but I guess we'll find out in the coming weeks.